add kill player and defeat check

+++ include/server/map.h
@@ -27,12 +27,12 @@ public:
 	const std::vector<Entity*> & getDynamicEntities() const;
 
 	b2Vec2 & getSpawn();
-	int getMinPlayers();
+	size_t getMinPlayers();
 
 	~Map();
 
 private:
-	int minPlayers;
+	size_t minPlayers;
 	b2Vec2 spawn;
 
 	std::vector<Entity*> staticEntities;
